requirements.txt
torch>=2.1.0
torchvision>=0.16.0
matplotlib>=3.7.1
pandas>=2.1.0
tensorboard>=2.15.0
Pillow>=10.0.0
numpy>=1.26.0
fastapi~=0.99.1
scikit-learn~=1.5.0
tqdm~=4.67.1
pydantic~=1.10.10